ai-agent-skills/21-external-registries/21-rushikeshpol-skills/skill-eval/requirements.txt
langfuse>=2.0.0
anthropic>=0.40.0
pyyaml>=6.0
rich>=13.0
python-dotenv>=1.0
